Module shuffle_transaction_builder::framework

Conversion library between a structured representation of a Move script call (ScriptCall) and the standard BCS-compatible representation used in Diem transactions (Script).

This code was generated by compiling known Script interfaces (“ABIs”) with the tool transaction-builder-generator.

Enums

ScriptCall

Structured representation of a call into a known Move script.

ScriptFunctionCall

Structured representation of a call into a known Move script function.

Functions

encode_add_currency_to_account_script_function

Summary

encode_burn_with_amount_script_function

Summary

encode_cancel_burn_with_amount_script_function

Summary

encode_create_designated_dealer_script_function

Summary

encode_create_regular_account_script_function

Create a regular account

encode_create_validator_account_script_function

Summary

encode_create_validator_operator_account_script_function

Summary

encode_freeze_account_script_function

Summary

encode_mint_coin_script_function

encode_preburn_script_function

Summary

encode_rotate_authentication_key_script_function

Summary

encode_rotate_authentication_key_with_nonce_admin_script_function

Summary

encode_rotate_authentication_key_with_nonce_script_function

Summary

encode_rotate_dual_attestation_info_script_function

Summary

encode_tiered_mint_script_function

Summary

encode_unfreeze_account_script_function

Summary

encode_update_dual_attestation_limit_script_function

Summary

encode_update_exchange_rate_script_function

Summary

encode_update_minting_ability_script_function

Summary